fix: refactor RestaurantDetailView to make it much smaller
