+++ FoodieFootprints/Resources/Localizable.xcstrings
@@ -20,16 +20,6 @@
     "%lld ★" : {
 
     },
-    "%lld of %lld" : {
-      "localizations" : {
-        "en" : {
-          "stringUnit" : {
-            "state" : "new",
-            "value" : "%1$lld of %2$lld"
-          }
-        }
-      }
-    },
     "+%lld" : {
 
     },
@@ -840,6 +830,7 @@
       }
     },
     "Are you sure you want to delete this visit? This action cannot be undone." : {
+      "extractionState" : "stale",
       "localizations" : {
         "de" : {
           "stringUnit" : {
